Move IntelliJ Plugin to dev.testify package
