+++ Plugins/IntelliJ/gradle.properties
@@ -1,7 +1,7 @@
 # IntelliJ Platform Artifacts Repositories
 # -> https://www.jetbrains.org/intellij/sdk/docs/reference_guide/intellij_artifacts.html
 
-pluginGroup = com.shopify.testify
+pluginGroup = dev.testify
 pluginName = Android Testify - Screenshot Instrumentation Tests
 pluginVersion = 0.1.2
 pluginSinceBuild = 193
