--- Plugins/IntelliJ/src/main/kotlin/com/shopify/testify/extensions/TooltipProvider.kt
@@ -1,8 +1,0 @@
-package com.shopify.testify.extensions
-
-import com.intellij.psi.PsiElement
-import com.intellij.util.Function
-
-typealias TooltipProvider = Function<PsiElement?, String?>
-
-const val SCREENSHOT_INSTRUMENTATION = "com.shopify.testify.annotation.ScreenshotInstrumentation"
