+++ Plugins/IntelliJ/src/main/resources/META-INF/plugin.xml
@@ -1,7 +1,7 @@
 <idea-plugin require-restart="false">
-    <id>com.shopify.testify</id>
+    <id>dev.testify</id>
     <name>Android Testify - Screenshot Instrumentation Tests</name>
-    <vendor email="support@shopify.com" url="http://testify.dev">Shopify</vendor>
+    <vendor email="testifyscreenshots@gmail.com" url="http://testify.dev">ndtp</vendor>
 
     <description><![CDATA[
 Testify screenshot tests are built on top of Android Instrumentation tests and so already integrate seamlessly with existing test suites. Screenshots can be captured directly from within Android Studio or using the Gradle command-line tools.
@@ -17,11 +17,11 @@ However, the current Android Studio support relies fully on the fact that Testif
     <extensions defaultExtensionNs="com.intellij">
         <codeInsight.lineMarkerProvider
                 language="kotlin"
-                implementationClass="com.shopify.testify.extensions.ScreenshotInstrumentationLineMarkerProvider"/>
+                implementationClass="dev.testify.extensions.ScreenshotInstrumentationLineMarkerProvider"/>
 
         <codeInsight.lineMarkerProvider
                 language="kotlin"
-                implementationClass="com.shopify.testify.extensions.ScreenshotClassMarkerProvider"/>
+                implementationClass="dev.testify.extensions.ScreenshotClassMarkerProvider"/>
 
     </extensions>
